refactor(sdk): Move TimelineInner into its own module
